Added API to MadLibsAPI
